third_party/blink/web_tests/platform/mac/external/wpt/editing/run/bold_1-1000-expected.txt

This is a testharness.js-based test.
Found 162 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["bold",""]] "foo[]bar" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "foo[]bar" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<p>[foo</p> <p>bar]</p>" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["bold",""]] "<p>[foo</p> <p>bar]</p>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<p>[foo</p> <p>bar]</p>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<p>[foo</p> <p>bar]</p>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<p>[foo</p> <p>bar]</p>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span>[foo</span> <span>bar]</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span>[foo</span> <span>bar]</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span>[foo</span> <span>bar]</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span>[foo</span> <span>bar]</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"font-weight:bold\\">foo</span></p><p> <span style=\\"font-weight:bold\\"><span>bar</span></span> </p><p><span style=\\"font-weight:bold\\">baz</span></p>" but got "<p><span style=\\"font-weight:bold\\">foo</span></p><p> <span style=\\"font-weight:bold\\"><span>bar</span> </span></p><p><span style=\\"font-weight:bold\\">baz</span></p>"
[FAIL] [["stylewithcss","true"],["bold",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><b>foo</b></p><p> <b><span>bar</span></b> </p><p><b>baz</b></p>" but got "<p><b>foo</b></p><p> <b><span>bar</span> </b></p><p><b>baz</b></p>"
[FAIL] [["stylewithcss","false"],["bold",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<p>[foo<p><br><p>bar]" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<p>[foo<p><br><p>bar]" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<p>[foo<p><br><p>bar]" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<p>[foo<p><br><p>bar]" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""]] "<b>foo[]bar</b>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""]] "<b>foo[]bar</b>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "<i>foo[]bar</i>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "<i>foo[]bar</i>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""]] "<span>foo</span>{}<span>bar</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "<span>foo</span>{}<span>bar</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""]] "<span>foo[</span><span>]bar</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "<span>foo[</span><span>]bar</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo[bar]baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo[bar]baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo[bar]baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo[bar]baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""]] "foo[bar<b>baz]qoz</b>quz" queryCommandIndeterm("bold") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["bold",""]] "foo[bar<b>baz]qoz</b>quz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "foo[bar<b>baz]qoz</b>quz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo[bar<i>baz]qoz</i>quz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo[bar<i>baz]qoz</i>quz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo[bar<i>baz]qoz</i>quz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo[bar<i>baz]qoz</i>quz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "{<p><p> <p>foo</p>}" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "{<p><p> <p>foo</p>}" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "{<p><p> <p>foo</p>}" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "{<p><p> <p>foo</p>}" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""]] "foo<span contenteditable=false>[bar]</span>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "foo<span contenteditable=false>[bar]</span>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "fo[o<span contenteditable=false>bar</span>b]az" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "fo[o<span contenteditable=false>bar</span>b]az" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "fo[o<span contenteditable=false>bar</span>b]az" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "fo[o<span contenteditable=false>bar</span>b]az" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""]] "foo<span contenteditable=false>ba[r</span>b]az" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "foo<span contenteditable=false>ba[r</span>b]az" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "fo[o<span contenteditable=false>b]ar</span>baz": execCommand("bold", false, "") return value
  assert_equals: expected false but got true
[FAIL] [["bold",""]] "fo[o<span contenteditable=false>b]ar</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span contenteditable=\\"false\\">bar</span>baz" but got "fo<b>o</b><span contenteditable=\\"false\\">bar</span>baz"
[FAIL] [["bold",""]] "fo[o<span contenteditable=false>b]ar</span>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "fo[o<span contenteditable=false>b]ar</span>baz" queryCommandState("bold") after
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["bold",""]] "fo[o<span contenteditable=false>b]ar</span>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "fo[<b>o</b><span contenteditable=false>bar</span><b>b</b>]az" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "fo[<b>o</b><span contenteditable=false>bar</span><b>b</b>]az" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "fo[<b>o</b><span contenteditable=false>bar</span><b>b</b>]az" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "fo[<b>o</b><span contenteditable=false>bar</span><b>b</b>]az" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span contenteditable=false>foo<span contenteditable=true>[bar]</span>baz</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span contenteditable=false>foo<span contenteditable=true>[bar]</span>baz</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span contenteditable=false>foo<span contenteditable=true>[bar]</span>baz</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span contenteditable=false>foo<span contenteditable=true>[bar]</span>baz</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""]] "<span contenteditable=false>fo[o<span contenteditable=true>bar</span>b]az</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "<span contenteditable=false>fo[o<span contenteditable=true>bar</span>b]az</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "<span contenteditable=false>foo<span contenteditable=true>ba[r</span>b]az</span>": execCommand("bold", false, "") return value
  assert_equals: expected false but got true
[FAIL] [["bold",""]] "<span contenteditable=false>foo<span contenteditable=true>ba[r</span>b]az</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span contenteditable=\\"false\\">foo<span contenteditable=\\"true\\">bar</span>baz</span>" but got "<span contenteditable=\\"false\\">foo<span contenteditable=\\"true\\">ba<b>r</b></span>baz</span>"
[FAIL] [["bold",""]] "<span contenteditable=false>foo<span contenteditable=true>ba[r</span>b]az</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "<span contenteditable=false>foo<span contenteditable=true>ba[r</span>b]az</span>" queryCommandState("bold") after
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["bold",""]] "<span contenteditable=false>foo<span contenteditable=true>ba[r</span>b]az</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""]] "<span contenteditable=false>fo[o<span contenteditable=true>b]ar</span>baz</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "<span contenteditable=false>fo[o<span contenteditable=true>b]ar</span>baz</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "<span contenteditable=false>fo[<b>o<span contenteditable=true>bar</span>b</b>]az</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""]] "<span contenteditable=false>fo[<b>o<span contenteditable=true>bar</span>b</b>]az</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<table><tbody><tr><td>foo<td>b[a]r<td>baz</table>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<table><tbody><tr><td>foo<td>b[a]r<td>baz</table>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<table><tbody><tr><td>foo<td>b[a]r<td>baz</table>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<table><tbody><tr><td>foo<td>b[a]r<td>baz</table>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<table><tbody><tr data-start=1 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<table><tbody><tr data-start=1 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<table><tbody><tr data-start=1 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<table><tbody><tr data-start=1 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<table><tbody><tr data-start=0 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<table><tbody><tr data-start=0 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<table><tbody><tr data-start=0 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<table><tbody><tr data-start=0 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandState("bold") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["bold",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandState("bold") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["bold",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandState("bold") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["bold",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandState("bold") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["bold",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<span style=\\"font-weight: bold\\">[bar]</span>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<span style=\\"font-weight: bold\\">[bar]</span>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo<span style=\\"font-weight: bold\\">[bar]</span>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo<span style=\\"font-weight: bold\\">[bar]</span>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<b>[bar]</b>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<b>[bar]</b>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo<b>[bar]</b>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo<b>[bar]</b>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "foo<b>bar</b>[baz]" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "foo<b>bar</b>[baz]" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""]] "[foo]<b>bar</b>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "[foo]<b>bar</b>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<b>foo</b>[bar]<b>baz</b>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<b>foo</b>[bar]<b>baz</b>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<b>foo</b>[bar]<b>baz</b>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<b>foo</b>[bar]<b>baz</b>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""]] "foo<strong>bar</strong>[baz]" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "foo<strong>bar</strong>[baz]" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""]] "[foo]<strong>bar</strong>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "[foo]<strong>bar</strong>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""]] "<strong>foo</strong>[bar]<strong>baz</strong>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "<strong>foo</strong>[bar]<strong>baz</strong>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""]] "<b>foo</b>[bar]<strong>baz</strong>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "<b>foo</b>[bar]<strong>baz</strong>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<strong>foo</strong>[bar]<b>baz</b>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<strong>foo</strong>[bar]<b>baz</b>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<strong>foo</strong>[bar]<b>baz</b>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<strong>foo</strong>[bar]<b>baz</b>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo[<b>bar</b>]baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo[<b>bar</b>]baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo[<b>bar</b>]baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo[<b>bar</b>]baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo[<b>bar]</b>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo[<b>bar]</b>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo[<b>bar]</b>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo[<b>bar]</b>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<b>[bar</b>]baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<b>[bar</b>]baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo<b>[bar</b>]baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo<b>[bar</b>]baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "foo{<b></b>}baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "foo{<b></b>}baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""]] "foo{<i></i>}baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "foo{<i></i>}baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""]] "foo{<b><i></i></b>}baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "foo{<b><i></i></b>}baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""]] "foo{<i><b></b></i>}baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "foo{<i><b></b></i>}baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""]] "foo<strong>[bar]</strong>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""]] "foo<strong>[bar]</strong>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "foo[<strong>bar</strong>]baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""]] "foo[<strong>bar</strong>]baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "foo[<strong>bar]</strong>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""]] "foo[<strong>bar]</strong>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "foo<strong>[bar</strong>]baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""]] "foo<strong>[bar</strong>]baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo[<span style=\\"font-weight: bold\\">bar</span>]baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo[<span style=\\"font-weight: bold\\">bar</span>]baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
Harness: the test ran to completion.
